LayoutTests/platform/wpe/imported/w3c/web-platform-tests/css/css-tables/border-writing-mode-dynamic-001-expected.txt

TABLE
TBODY
TR
TD

FAIL Table borders track writing mode changes assert_equals: expected 132 but got 152
